% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile-init.R
\name{.init_pdata}
\alias{.init_pdata}
\title{Preps the pdata(x) from a bioconductor container to be a legit facile-looking
sample frame}
\usage{
.init_pdata(x, sinfo = pdata(x), ...)
}
\arguments{
\item{x}{a bioconductor container}

\item{sinfo}{the data.frame from \code{pdata(x)}}
}
\description{
This function is not exported on purpose.
}
