% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_bioc_data.R
\name{example_bioc_data}
\alias{example_bioc_data}
\title{Create an example Bioconductor assay container}
\usage{
example_bioc_data(class = "DGEList", efds = NULL, y = NULL, ...)
}
\arguments{
\item{type}{The type of assay container. Pick from the ones enumerated above.}
}
\value{
The Bioconductor container
}
\description{
We can construct the following types of containers:
}
\details{
\enumerate{
\item DGEList
\item EList
\item ExpressionSet
\item SummarizedExperiment
\item DESeqDataSet
\item SingleCellExperiment (not yet)
\item MultiAssayExperiment (not yet)
}
}
