% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEList.R, R/ExpressionSet.R,
%   R/MultiAssayExperiment.R, R/SingleCellExperiment.R,
%   R/SummarizedExperiment.R, R/api.R
\name{facilitate.DGEList}
\alias{facilitate.DGEList}
\alias{facilitate.ExpressionSet}
\alias{facilitate.MultiAssayExperiment}
\alias{facilitate.SingleCellExperiment}
\alias{facilitate.SummarizedExperiment}
\alias{facilitate}
\title{Creates FacileData sublcasses of Bioconductor assay containers}
\usage{
\method{facilitate}{DGEList}(x, ...)

\method{facilitate}{ExpressionSet}(x, ...)

\method{facilitate}{MultiAssayExperiment}(x, ...)

\method{facilitate}{SingleCellExperiment}(x, ...)

\method{facilitate}{SummarizedExperiment}(x, ...)

facilitate(x, ...)
}
\arguments{
\item{x}{A bioconductor assay container}

\item{...}{parameters to tweak container-specific funcitonality}
}
\value{
A facile-subclass of \code{x} that implements teh FacileData API
}
\description{
This function creates a subclass of given Bioconductor assay container that
implements the FacileData API. This is in contrast with the
\code{\link[FacileData:as.FacileDataSet]{FacileData::as.FacileDataSet()}} function, which literally converts an input
object into a completely new \code{FacileDataSet}.
}
