% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{FacileBiocData-package}
\alias{FacileBiocData}
\alias{FacileBiocData-package}
\title{FacileBiocData: FacileData API wrappers over Bioconductor assay containers}
\description{
Enables the use of standard Bioconductor assay containers, like
    SummarizedExperiment, MultiAssayExperiment, DGEList, ExpressionSet, etc.
    within the facile.bio ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/facilebio/FacileBiocData}
  \item Report bugs at \url{https://github.com/facilebio/FacileBiocData/issues}
}

}
\author{
\strong{Maintainer}: Steve Lianoglou \email{lianoglou@dnli.com}

Other contributors:
\itemize{
  \item Denali Therapeutics [copyright holder, funder]
}

}
