% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{class}
\name{FacileBiocDataStore-class}
\alias{FacileBiocDataStore-class}
\title{Root virtual class that signals object as a FacileDataStore}
\description{
All FacileBiocDataStore objects should inherit from this
}
